+++ doc/2/controllers/auth/update-my-credentials/snippets/update-my-credentials.dart
@@ -1,0 +1,19 @@
+await kuzzle
+  .auth
+  .login(
+    'local',
+    {
+      'username': 'user',
+      'password': 'bar'
+    },
+);
+
+final result = await kuzzle
+  .auth
+  .updateMyCredentials(
+    'local',
+    {
+      'username': 'user2',
+      'password': 'bar2'
+    }
+);