+++ doc/2/controllers/security/m-delete-users/snippets/m-delete-users.dart
@@ -1,0 +1,5 @@
+final result = await kuzzle.security.mDeleteUsers([
+  'mDeleteUser1',
+  'mDeleteUser2',
+  'mDeleteUser3'
+]);