+++ lib/kuzzle.dart
@@ -1,4 +1,4 @@
-library kuzzle_dart;
+library sdk_kuzzle;
 
 export 'src/controllers/abstract.dart';
 export 'src/controllers/auth.dart';
@@ -15,13 +15,12 @@ export 'src/kuzzle/profile.dart';
 export 'src/kuzzle/request.dart';
 export 'src/kuzzle/response.dart';
 export 'src/kuzzle/role.dart';
-export 'src/kuzzle/room.dart';
 export 'src/kuzzle/user.dart';
 export 'src/protocols/abstract.dart';
 export 'src/protocols/websocket.dart';
-export 'src/search_result/abstract.dart';
 export 'src/search_result/documents.dart';
 export 'src/search_result/profiles.dart';
 export 'src/search_result/roles.dart';
+export 'src/search_result/search-result.dart';
 export 'src/search_result/specifications.dart';
 export 'src/search_result/users.dart';