+++ lib/src/search_result/users.dart
@@ -3,15 +3,14 @@ import '../kuzzle/request.dart';
 import '../kuzzle/response.dart';
 import '../kuzzle/user.dart';
 
-import 'abstract.dart';
+import 'search-result.dart';
 
-class UserSearchResult extends KuzzleSearchResult {
+class UserSearchResult extends SearchResult {
   UserSearchResult(
     Kuzzle kuzzle, {
     KuzzleRequest request,
     KuzzleResponse response,
   }) : super(kuzzle, request: request, response: response) {
-    controller = null;
     searchAction = 'searchUsers';
     scrollAction = 'scrollUsers';
 
@@ -24,11 +23,14 @@ class UserSearchResult extends KuzzleSearchResult {
   }
 
   @override
-  Future<List<dynamic>> next() => super.next().then((_) => hits =
-      (response.result['hits'] as List).map((hit) => KuzzleUser(kuzzle,
-          uid: hit['_id'] as String,
-          content: hit['_source'] as Map<String, dynamic>,
-          meta: hit['_meta'] as Map<String, dynamic>)) as List<dynamic>);
+  UserSearchResult buildNextSearchResult (KuzzleResponse response) {
+    final nextSearchResult = UserSearchResult(
+      kuzzle, 
+      request: request, 
+      response: response);
+    nextSearchResult.fetched += fetched;
+    return nextSearchResult;
+  }
 
   List<KuzzleUser> getUsers() => List<KuzzleUser>.from(hits);
 }