+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [2.1.0]
+
+- Fix KuzzleError instanciation (Yoann-Abbes)
+- Add security:searchApiKeys (Yoann-Abbes)
+- Able to use "SearchResult.next()" for securities (Yoann-Abbes)
+- Fix a concurrent modification error during reconnection (#44) (Yoann-Abbes)
+- Add support for the security:create and security:deleteApiKey API actions (Yoann-Abbes)
+- Add security:refresh and security:mGetUsers (Yoann-Abbes)
+
+## [2.0.1]
+
+- Fix README.md
+- Deleted examples
+
 ## [2.0.0-alpha.1] : 26.03.2020
 
 **Breaking Changes**